Always inline opPre
opPre is marked as inline, but our current gcc version isn't
actually inlining it anywhere.  I find that surprising, because it
seems like a no-brainer: the function is small and the code gets
better when inlined (e.g., it gets rid of stores to reference params).
This diff makes the text very slightly bigger but should have good
effects on locality within the helpers.