Adds a testcase for another SimpleXML bug.
Even nodes that only contain text are considered an "array-like" node
with the text being on index 0. This should not be the case.

This bug was somewhat solved by a pull-request that was never merged,
but I question the way it was implemented. See the following commit:
https://github.com/facebook/hiphop-php/commit/d35d81360eb0db5abbd5ad601fde94255e30fbd3

Input on how to solve this is very welcome.