Adds a testcase for another SimpleXML bug.
Even nodes that only contain text are considered an "array-like" node
with the text being on index 0.

This bug was somewhat solved by a pull-request that was never merged,
but I question the way it was implemented. See the following commit:
https://github.com/facebook/hiphop-php/commit/d35d81360eb0db5abbd5ad601fde94255e30fbd3

I could very well try to merge the above commit into current code, but
would like the input of hphp maintainers on how this should be
implemented.

I'm willing to commit to this project in the long-term, but some kind of
mentorship deal would benefit both of us greatly :)