import zend pdo_mysql tests
I did some minor surgery to the import script to always copy in the php files. This is because zend has ##config.phpt## files which, as their name doens't reveal, aren't really tests but instead are config files.

After all that, the tests still didn't work because I'm not running mysql on localhost.