fix race condition in SplFileInfo_getInode_basic
These failed once on @smith's diff and I think it was from them both running at the same time.

I'm not really sure why the others changed their path, but that is the original zend file.
It shouldn't matter as it is a error line that will be set to whatever we print out if the test passes.

Done with:

  tools/import_zend_test.py -z /tmp/php-5.4.13/ -o SplFileInfo_get