fix shared test name
This test was imported WITHOUT hard-coded output. Meaning it was compared to zend. It so happened that I had cleaned my repo after any runs of ##TestExt## so the file they both created (and sometimes didn't clean up) didn't exist. So that is the output I recorded.

@mwilliams do you think you intended for the file to exist? It wasn't in the repo.

With this diff I'm using a unique filename so it doesn't matter if ##TestExt## hasn't cleaned up perfectly.